Fix BLOB migration when LinguaPlone is installed.
When LinguaPlone is installed, translations are updated for language
independent fields, some of which may not have been migrated yet.
Since the migrator was passing in an actual content object to the
mutator but only the ATBlob mutator could handle that and not the old
ATImage mutator, this was causing the migration to fail for individual
images that had translations.

A further side-effect of this exposes a potential ZODB bug.  When the
migration fails for an individual image, the savepoint is rolled
back.  When the transaction is then committed for those images the
transaction succeeded for a POSKeyError can happen.